+++ tests/evaluation_proof.rs
@@ -8,7 +8,7 @@ use dory::{
     curve::commit_and_evaluate_batch,
     setup::ProverSetup,
     toy_transcript::ToyTranscript,
-    core::pcs::{create_evaluation_proof, verify_evaluation_proof},
+    core::evaluate::{create_evaluation_proof, verify_evaluation_proof},
 };
 
 use dory::curve::{
@@ -21,8 +21,8 @@ fn test_evaluation_proof() {
     let total_start = Instant::now();
 
     // ----- Test Parameters -----
-    let length: usize = 1 << 12;
-    let max_log_n: usize = 12;
+    let length: usize = 1 << 9;
+    let max_log_n: usize = 9;
 
     println!("Parameters:");
     println!("  - Polynomial length: {}", length);
@@ -408,7 +408,7 @@ fn test_evaluation_proof_tampered_messages_should_fail() {
     // Create fresh transcript for verification
     let verify_transcript = ToyTranscript::new(domain);
 
-    let verification_result = dory::core::pcs::verify_evaluation_proof::<
+    let verification_result = dory::core::evaluate::verify_evaluation_proof::<
         ArkBn254Pairing,
         ToyTranscript,
         OptimizedMsmG1,
